+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**3.3.1**
 
 - Added missing documentation for the ``ExceptionGroup.exceptions`` attribute
 - Changed the asyncio test runner not to use uvloop by default (to match the behavior of